Botons amb command

+++ Challenge1_2/Common/Converters/ZeroToPlaceholderConverter.cs
@@ -8,7 +8,7 @@ namespace Challenge1_2.Common.Converters
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (System.Convert.ToInt16(value) == 0) ? " " : value;
+            return (System.Convert.ToInt32(value) == 0) ? " " : value;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
